Analyse Dataset with the Cox Protportional Hazards Model

Analyse Dataset with the Cox Protportional Hazards Model

Usage

analyse_coxph(level = 0.95, alternative = "two.sided")

Arguments

level

confidence level for CI computation

alternative

alternative hypothesis for the tests "two.sided" or "one.sieded"

Value

an analyse function that returns a list with the elements

p p value of the score test (two.sided) or the Wald test (one.sided)

alternative the alternative used

coef coefficient for trt

hr hazard ratio for trt

hr_lower lower 95% confidence intervall boundary for the hazard ratio for trt

hr_upperlower 95% confidence intervall boundary for the hazard ratio for trt

CI_level the CI level used

N_pat number of patients

N_evt number of events

Details

alternative can be "two.sided" for a two sided test of equality of the summary statistic or "one.sided" for a one sided test testing H0: treatment has equal or shorter survival than control vs. H1 treatment has longer survival than control.

Examples

condition <- merge(
  assumptions_delayed_effect(),
  design_fixed_followup(),
  by = NULL
) |>
  head(1)
#> expand.grid(
#>   delay=m2d(seq(0, 10, by=2)), # delay of 0, 1, ..., 10 months
#>   hazard_ctrl=m2r(24),         # median survival control of 24 months
#>   hazard_trt=m2r(36),          # median survival treatment of 36 months
#>   random_withdrawal=m2r(120)   # median time to random withdrawal 10 years
#> )
#> expand.grid(
#>   n_trt=150,         # 150 patients in the treatment arm
#>   n_ctrl=150,        # 150 patients in the control arm
#>   followup=m2d(24),  # followup 2 years
#>   recruitment=m2d(6) # recruitment time 6 months
#> 
#> )
dat <- generate_delayed_effect(condition)
analyse_coxph()(condition, dat)
#> $p
#> [1] 4.082247e-06
#> 
#> $alternative
#> [1] "two.sided"
#> 
#> $coef
#>        trt 
#> -0.5553412 
#> 
#> $hr
#>       trt 
#> 0.5738764 
#> 
#> $hr_lower
#> [1] 0.4518774
#> 
#> $hr_upper
#> [1] 0.728813
#> 
#> $CI_level
#> [1] 0.95
#> 
#> $N_pat
#> [1] 300
#> 
#> $N_evt
#> [1] 300
#>