Fast C implementation of hazard, cumulative hazard, ... for piecewise constant hazards

Fast C implementation of hazard, cumulative hazard, ... for piecewise constant hazards

Usage

fast_haz_fun(Tint, lambda)

fast_cumhaz_fun(Tint, lambda)

fast_cdf_fun(Tint, lambda)

fast_pdf_fun(Tint, lambda)

fast_surv_fun(Tint, lambda)

fast_quant_fun(Tint, lambda)

fast_rng_fun(Tint, lambda, discrete = TRUE)

Arguments

Tint

left points of the time intervals the hazards are defined on

lambda

hazards in the time intervals

discrete

should surivial times be rounded down to the next whole day

Value

A function with one parameter, a vector of times where the function should be evaluated.

for fast_rng_fun: a function with the argument that draws n samples from the survival distribution

Details

the last time interval extends to +Inf

Functions

fast_haz_fun(): fast hazard function

fast_cumhaz_fun(): fast cumulative hazard function

fast_cdf_fun(): fast cumulative density function function

fast_pdf_fun(): fast probability density function

fast_surv_fun(): fast survival function

fast_quant_fun(): fast quantile function

fast_rng_fun(): fast random numbers from survival distribution

Examples

haz <- fast_haz_fun(c(0, 10, 20), c(10, 15, 7))
haz(seq(0, 30, by=0.1))
#>   [1] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#>  [26] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#>  [51] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#>  [76] 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10 10
#> [101] 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> [126] 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> [151] 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> [176] 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15 15
#> [201]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [226]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [251]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [276]  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7  7
#> [301]  7
cumhaz <- fast_cumhaz_fun(c(0, 10, 20), c(10, 15, 7))
cumhaz(seq(0, 30, by=0.1))
#>   [1]   0.0   1.0   2.0   3.0   4.0   5.0   6.0   7.0   8.0   9.0  10.0  11.0
#>  [13]  12.0  13.0  14.0  15.0  16.0  17.0  18.0  19.0  20.0  21.0  22.0  23.0
#>  [25]  24.0  25.0  26.0  27.0  28.0  29.0  30.0  31.0  32.0  33.0  34.0  35.0
#>  [37]  36.0  37.0  38.0  39.0  40.0  41.0  42.0  43.0  44.0  45.0  46.0  47.0
#>  [49]  48.0  49.0  50.0  51.0  52.0  53.0  54.0  55.0  56.0  57.0  58.0  59.0
#>  [61]  60.0  61.0  62.0  63.0  64.0  65.0  66.0  67.0  68.0  69.0  70.0  71.0
#>  [73]  72.0  73.0  74.0  75.0  76.0  77.0  78.0  79.0  80.0  81.0  82.0  83.0
#>  [85]  84.0  85.0  86.0  87.0  88.0  89.0  90.0  91.0  92.0  93.0  94.0  95.0
#>  [97]  96.0  97.0  98.0  99.0 100.0 101.5 103.0 104.5 106.0 107.5 109.0 110.5
#> [109] 112.0 113.5 115.0 116.5 118.0 119.5 121.0 122.5 124.0 125.5 127.0 128.5
#> [121] 130.0 131.5 133.0 134.5 136.0 137.5 139.0 140.5 142.0 143.5 145.0 146.5
#> [133] 148.0 149.5 151.0 152.5 154.0 155.5 157.0 158.5 160.0 161.5 163.0 164.5
#> [145] 166.0 167.5 169.0 170.5 172.0 173.5 175.0 176.5 178.0 179.5 181.0 182.5
#> [157] 184.0 185.5 187.0 188.5 190.0 191.5 193.0 194.5 196.0 197.5 199.0 200.5
#> [169] 202.0 203.5 205.0 206.5 208.0 209.5 211.0 212.5 214.0 215.5 217.0 218.5
#> [181] 220.0 221.5 223.0 224.5 226.0 227.5 229.0 230.5 232.0 233.5 235.0 236.5
#> [193] 238.0 239.5 241.0 242.5 244.0 245.5 247.0 248.5 250.0 250.7 251.4 252.1
#> [205] 252.8 253.5 254.2 254.9 255.6 256.3 257.0 257.7 258.4 259.1 259.8 260.5
#> [217] 261.2 261.9 262.6 263.3 264.0 264.7 265.4 266.1 266.8 267.5 268.2 268.9
#> [229] 269.6 270.3 271.0 271.7 272.4 273.1 273.8 274.5 275.2 275.9 276.6 277.3
#> [241] 278.0 278.7 279.4 280.1 280.8 281.5 282.2 282.9 283.6 284.3 285.0 285.7
#> [253] 286.4 287.1 287.8 288.5 289.2 289.9 290.6 291.3 292.0 292.7 293.4 294.1
#> [265] 294.8 295.5 296.2 296.9 297.6 298.3 299.0 299.7 300.4 301.1 301.8 302.5
#> [277] 303.2 303.9 304.6 305.3 306.0 306.7 307.4 308.1 308.8 309.5 310.2 310.9
#> [289] 311.6 312.3 313.0 313.7 314.4 315.1 315.8 316.5 317.2 317.9 318.6 319.3
#> [301] 320.0
cdf <- fast_cdf_fun(c(0, 10, 20), c(10, 15, 7))
cdf(seq(0, 30, by=0.1))
#>   [1] 0.0000000 0.6321206 0.8646647 0.9502129 0.9816844 0.9932621 0.9975212
#>   [8] 0.9990881 0.9996645 0.9998766 0.9999546 0.9999833 0.9999939 0.9999977
#>  [15] 0.9999992 0.9999997 0.9999999 1.0000000 1.0000000 1.0000000 1.0000000
#>  [22] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [29] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [36] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [43] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [50] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [57] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [64] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [71] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [78] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [85] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [92] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#>  [99] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [106] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [113] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [120] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [127] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [134] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [141] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [148] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [155] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [162] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [169] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [176] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [183] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [190] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [197] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [204] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [211] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [218] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [225] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [232] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [239] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [246] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [253] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [260] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [267] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [274] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [281] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [288] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
#> [295] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000
pdf <- fast_pdf_fun(c(0, 10, 20), c(10, 15, 7))
pdf(seq(0, 30, by=0.1))
#>   [1]  1.000000e+01  3.678794e+00  1.353353e+00  4.978707e-01  1.831564e-01
#>   [6]  6.737947e-02  2.478752e-02  9.118820e-03  3.354626e-03  1.234098e-03
#>  [11]  4.539993e-04  1.670170e-04  6.144212e-05  2.260329e-05  8.315287e-06
#>  [16]  3.059023e-06  1.125352e-06  4.139938e-07  1.522998e-07  5.602796e-08
#>  [21]  2.061154e-08  7.582560e-09  2.789468e-09  1.026188e-09  3.775135e-10
#>  [26]  1.388794e-10  5.109089e-11  1.879529e-11  6.914400e-12  2.543666e-12
#>  [31]  9.357623e-13  3.442477e-13  1.266417e-13  4.658886e-14  1.713908e-14
#>  [36]  6.305117e-15  2.319523e-15  8.533048e-16  3.139133e-16  1.154822e-16
#>  [41]  4.248354e-17  1.562882e-17  5.749522e-18  2.115131e-18  7.781132e-19
#>  [46]  2.862519e-19  1.053062e-19  3.873998e-20  1.425164e-20  5.242886e-21
#>  [51]  1.928750e-21  7.095474e-22  2.610279e-22  9.602680e-23  3.532629e-23
#>  [56]  1.299581e-23  4.780893e-24  1.758792e-24  6.470235e-25  2.380266e-25
#>  [61]  8.756511e-26  3.221340e-26  1.185065e-26  4.359610e-27  1.603811e-27
#>  [66]  5.900091e-28  2.170522e-28  7.984904e-29  2.937482e-29  1.080639e-29
#>  [71]  3.975450e-30  1.462486e-30  5.380186e-31  1.979260e-31  7.281290e-32
#>  [76]  2.678637e-32  9.854155e-33  3.625141e-33  1.333615e-33  4.906095e-34
#>  [81]  1.804851e-34  6.639677e-35  2.442601e-35  8.985826e-36  3.305701e-36
#>  [86]  1.216099e-36  4.473779e-37  1.645811e-37  6.054602e-38  2.227364e-38
#>  [91]  8.194013e-39  3.014409e-39  1.108939e-39  4.079559e-40  1.500786e-40
#>  [96]  5.521082e-41  2.031093e-41  7.471972e-42  2.748785e-42  1.011221e-42
#> [101]  5.580114e-43  1.245092e-43  2.778175e-44  6.198947e-45  1.383172e-45
#> [106]  3.086274e-46  6.886408e-47  1.536565e-47  3.428541e-48  7.650108e-49
#> [111]  1.706970e-49  3.808764e-50  8.498502e-51  1.896272e-51  4.231155e-52
#> [116]  9.440983e-53  2.106568e-53  4.700389e-54  1.048798e-54  2.340186e-55
#> [121]  5.221660e-56  1.165110e-56  2.599712e-57  5.800741e-58  1.294320e-58
#> [126]  2.888019e-59  6.444041e-60  1.437860e-60  3.208299e-61  7.158683e-62
#> [131]  1.597318e-62  3.564098e-63  7.952578e-64  1.774460e-64  3.959355e-65
#> [136]  8.834516e-66  1.971247e-66  4.398447e-67  9.814261e-68  2.189858e-68
#> [141]  4.886233e-69  1.090266e-69  2.432712e-70  5.428114e-71  1.211176e-71
#> [146]  2.702499e-72  6.030090e-73  1.345495e-73  3.002205e-74  6.698825e-75
#> [151]  1.494710e-75  3.335149e-76  7.441723e-77  1.660473e-77  3.705016e-78
#> [156]  8.267007e-79  1.844619e-79  4.115900e-80  9.183815e-81  2.049186e-81
#> [161]  4.572352e-82  1.020230e-82  2.276440e-83  5.079425e-84  1.133373e-84
#> [166]  2.528897e-85  5.642731e-86  1.259064e-86  2.809350e-87  6.268508e-88
#> [171]  1.398693e-88  3.120906e-89  6.963683e-90  1.553808e-90  3.467014e-91
#> [176]  7.735954e-92  1.726125e-92  3.851504e-93  8.593868e-94  1.917551e-94
#> [181]  4.278635e-95  9.546925e-96  2.130207e-96  4.753134e-97  1.060568e-97
#> [186]  2.366446e-98  5.280255e-99  1.178184e-99 2.628884e-100 5.865834e-101
#> [191] 1.308844e-101 2.920427e-102 6.516352e-103 1.453995e-103 3.244301e-104
#> [196] 7.239014e-105 1.615242e-105 3.604093e-106 8.041818e-107 1.794372e-107
#> [201] 1.868433e-108 9.278364e-109 4.607499e-109 2.288017e-109 1.136195e-109
#> [206] 5.642179e-110 2.801823e-110 1.391344e-110 6.909211e-111 3.431013e-111
#> [211] 1.703790e-111 8.460773e-112 4.201496e-112 2.086401e-112 1.036076e-112
#> [216] 5.145001e-113 2.554932e-113 1.268742e-113 6.300385e-114 3.128679e-114
#> [221] 1.553656e-114 7.715226e-115 3.831268e-115 1.902551e-115 9.447791e-116
#> [226] 4.691634e-116 2.329797e-116 1.156943e-116 5.745208e-117 2.852986e-117
#> [231] 1.416751e-117 7.035376e-118 3.493664e-118 1.734902e-118 8.615270e-119
#> [236] 4.278216e-119 2.124499e-119 1.054995e-119 5.238951e-120 2.601586e-120
#> [241] 1.291909e-120 6.415432e-121 3.185809e-121 1.582026e-121 7.856109e-122
#> [246] 3.901228e-122 1.937293e-122 9.620311e-123 4.777305e-123 2.372339e-123
#> [251] 1.178069e-123 5.850117e-124 2.905082e-124 1.442621e-124 7.163844e-125
#> [256] 3.557460e-125 1.766582e-125 8.772588e-126 4.356338e-126 2.163294e-126
#> [261] 1.074260e-126 5.334616e-127 2.649092e-127 1.315500e-127 6.532581e-128
#> [266] 3.243984e-128 1.610915e-128 7.999565e-129 3.972466e-129 1.972668e-129
#> [271] 9.795981e-130 4.864540e-130 2.415659e-130 1.199581e-130 5.956942e-131
#> [276] 2.958130e-131 1.468964e-131 7.294659e-132 3.622420e-132 1.798841e-132
#> [281] 8.932779e-133 4.435887e-133 2.202796e-133 1.093876e-133 5.432028e-134
#> [286] 2.697465e-134 1.339522e-134 6.651868e-135 3.303220e-135 1.640330e-135
#> [291] 8.145640e-136 4.045005e-136 2.008690e-136 9.974860e-137 4.953369e-137
#> [296] 2.459770e-137 1.221486e-137 6.065718e-138 3.012147e-138 1.495788e-138
#> [301] 7.427862e-139
surv <- fast_surv_fun(c(0, 10, 20), c(10, 15, 7))
surv(seq(0, 30, by=0.1))
#>   [1]  1.000000e+00  3.678794e-01  1.353353e-01  4.978707e-02  1.831564e-02
#>   [6]  6.737947e-03  2.478752e-03  9.118820e-04  3.354626e-04  1.234098e-04
#>  [11]  4.539993e-05  1.670170e-05  6.144212e-06  2.260329e-06  8.315287e-07
#>  [16]  3.059023e-07  1.125352e-07  4.139938e-08  1.522998e-08  5.602796e-09
#>  [21]  2.061154e-09  7.582560e-10  2.789468e-10  1.026188e-10  3.775135e-11
#>  [26]  1.388794e-11  5.109089e-12  1.879529e-12  6.914400e-13  2.543666e-13
#>  [31]  9.357623e-14  3.442477e-14  1.266417e-14  4.658886e-15  1.713908e-15
#>  [36]  6.305117e-16  2.319523e-16  8.533048e-17  3.139133e-17  1.154822e-17
#>  [41]  4.248354e-18  1.562882e-18  5.749522e-19  2.115131e-19  7.781132e-20
#>  [46]  2.862519e-20  1.053062e-20  3.873998e-21  1.425164e-21  5.242886e-22
#>  [51]  1.928750e-22  7.095474e-23  2.610279e-23  9.602680e-24  3.532629e-24
#>  [56]  1.299581e-24  4.780893e-25  1.758792e-25  6.470235e-26  2.380266e-26
#>  [61]  8.756511e-27  3.221340e-27  1.185065e-27  4.359610e-28  1.603811e-28
#>  [66]  5.900091e-29  2.170522e-29  7.984904e-30  2.937482e-30  1.080639e-30
#>  [71]  3.975450e-31  1.462486e-31  5.380186e-32  1.979260e-32  7.281290e-33
#>  [76]  2.678637e-33  9.854155e-34  3.625141e-34  1.333615e-34  4.906095e-35
#>  [81]  1.804851e-35  6.639677e-36  2.442601e-36  8.985826e-37  3.305701e-37
#>  [86]  1.216099e-37  4.473779e-38  1.645811e-38  6.054602e-39  2.227364e-39
#>  [91]  8.194013e-40  3.014409e-40  1.108939e-40  4.079559e-41  1.500786e-41
#>  [96]  5.521082e-42  2.031093e-42  7.471972e-43  2.748785e-43  1.011221e-43
#> [101]  3.720076e-44  8.300611e-45  1.852117e-45  4.132631e-46  9.221146e-47
#> [106]  2.057516e-47  4.590938e-48  1.024377e-48  2.285694e-49  5.100072e-50
#> [111]  1.137980e-50  2.539176e-51  5.665668e-52  1.264181e-52  2.820770e-53
#> [116]  6.293989e-54  1.404379e-54  3.133593e-55  6.991990e-56  1.560124e-56
#> [121]  3.481107e-57  7.767399e-58  1.733141e-58  3.867160e-59  8.628801e-60
#> [126]  1.925346e-60  4.296027e-61  9.585732e-62  2.138866e-62  4.772455e-63
#> [131]  1.064879e-63  2.376065e-64  5.301719e-65  1.182973e-65  2.639570e-66
#> [136]  5.889677e-67  1.314165e-67  2.932298e-68  6.542841e-69  1.459905e-69
#> [141]  3.257489e-70  7.268439e-71  1.621808e-71  3.618743e-72  8.074507e-73
#> [146]  1.801666e-73  4.020060e-74  8.969967e-75  2.001470e-75  4.465884e-76
#> [151]  9.964733e-77  2.223432e-77  4.961148e-78  1.106982e-78  2.470010e-79
#> [156]  5.511338e-80  1.229746e-80  2.743934e-81  6.122544e-82  1.366124e-82
#> [161]  3.048235e-83  6.801532e-84  1.517627e-84  3.386283e-85  7.555819e-86
#> [166]  1.685931e-86  3.761821e-87  8.393757e-88  1.872900e-88  4.179005e-89
#> [171]  9.324621e-90  2.080604e-90  4.642456e-91  1.035872e-91  2.311343e-92
#> [176]  5.157302e-93  1.150750e-93  2.567670e-94  5.729245e-95  1.278367e-95
#> [181]  2.852423e-96  6.364617e-97  1.420138e-97  3.168756e-98  7.070451e-99
#> [186]  1.577631e-99 3.520170e-100 7.854561e-101 1.752589e-101 3.910556e-102
#> [191] 8.725629e-103 1.946951e-103 4.344235e-104 9.693298e-105 2.162867e-105
#> [196] 4.826009e-106 1.076828e-106 2.402728e-107 5.361212e-108 1.196248e-108
#> [201] 2.669190e-109 1.325481e-109 6.582142e-110 3.268595e-110 1.623136e-110
#> [206] 8.060256e-111 4.002605e-111 1.987635e-111 9.870302e-112 4.901447e-112
#> [211] 2.433986e-112 1.208682e-112 6.002137e-113 2.980573e-113 1.480109e-113
#> [216] 7.350002e-114 3.649903e-114 1.812488e-114 9.000550e-115 4.469541e-115
#> [221] 2.219508e-115 1.102175e-115 5.473240e-116 2.717931e-116 1.349684e-116
#> [226] 6.702334e-117 3.328281e-117 1.652775e-117 8.207439e-118 4.075694e-118
#> [231] 2.023930e-118 1.005054e-118 4.990949e-119 2.478432e-119 1.230753e-119
#> [236] 6.111738e-120 3.034999e-120 1.507136e-120 7.484216e-121 3.716552e-121
#> [241] 1.845585e-121 9.164903e-122 4.551156e-122 2.260037e-122 1.122301e-122
#> [246] 5.573184e-123 2.767561e-123 1.374330e-123 6.824721e-124 3.389056e-124
#> [251] 1.682956e-124 8.357310e-125 4.150117e-125 2.060887e-125 1.023406e-125
#> [256] 5.082086e-126 2.523689e-126 1.253227e-126 6.223340e-127 3.090419e-127
#> [261] 1.534657e-127 7.620880e-128 3.784417e-128 1.879286e-128 9.332258e-129
#> [266] 4.634262e-129 2.301306e-129 1.142795e-129 5.674952e-130 2.818098e-130
#> [271] 1.399426e-130 6.949343e-131 3.450942e-131 1.713687e-131 8.509918e-132
#> [276] 4.225900e-132 2.098520e-132 1.042094e-132 5.174886e-133 2.569772e-133
#> [281] 1.276111e-133 6.336981e-134 3.146852e-134 1.562680e-134 7.760041e-135
#> [286] 3.853522e-135 1.913602e-135 9.502668e-136 4.718885e-136 2.343329e-136
#> [291] 1.163663e-136 5.778579e-137 2.869557e-137 1.424980e-137 7.076241e-138
#> [296] 3.513957e-138 1.744980e-138 8.665312e-139 4.303067e-139 2.136840e-139
#> [301] 1.061123e-139
quant <- fast_quant_fun(c(0, 10, 20), c(10, 15, 7))
quant(seq(0, 1, by=0.01))
#>   [1] 0.000000000 0.001005034 0.002020271 0.003045921 0.004082199 0.005129329
#>   [7] 0.006187540 0.007257069 0.008338161 0.009431068 0.010536052 0.011653382
#>  [13] 0.012783337 0.013926207 0.015082289 0.016251893 0.017435339 0.018632958
#>  [19] 0.019845094 0.021072103 0.022314355 0.023572233 0.024846136 0.026136476
#>  [25] 0.027443685 0.028768207 0.030110509 0.031471074 0.032850407 0.034249031
#>  [31] 0.035667494 0.037106368 0.038566248 0.040047757 0.041551544 0.043078292
#>  [37] 0.044628710 0.046203546 0.047803580 0.049429632 0.051082562 0.052763274
#>  [43] 0.054472718 0.056211892 0.057981850 0.059783700 0.061618614 0.063487827
#>  [49] 0.065392647 0.067334455 0.069314718 0.071334989 0.073396918 0.075502258
#>  [55] 0.077652879 0.079850770 0.082098055 0.084397007 0.086750057 0.089159812
#>  [61] 0.091629073 0.094160854 0.096758403 0.099425227 0.102165125 0.104982212
#>  [67] 0.107880966 0.110866262 0.113943428 0.117118298 0.120397280 0.123787436
#>  [73] 0.127296568 0.130933332 0.134707365 0.138629436 0.142711636 0.146967597
#>  [79] 0.151412773 0.156064775 0.160943791 0.166073121 0.171479843 0.177195684
#>  [85] 0.183258146 0.189711998 0.196611286 0.204022083 0.212026354 0.220727491
#>  [91] 0.230258509 0.240794561 0.252572864 0.265926004 0.281341072 0.299573227
#>  [97] 0.321887582 0.350655790 0.391202301 0.460517019         NaN
rng <- fast_rng_fun(c(0, 10, 20), c(10, 15, 7))
rng(100)
#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1