Merge results from additional or updated simulations

Merge results from additional or updated simulations

Usage

upsert_merge(x, y, by)

merge_additional_results(
  old,
  new,
  design_names = NULL,
  descriptive_regex = NULL
)

Arguments

x

left data.frame

y

right data.frame

by

columns to match by

old

old results

new

new/additional results

design_names

names of the paramterst

descriptive_regex

regular expression for columns of descriptive statistics

Value

a data.frame

a data.frame of the merged simulation results

Details

updates columns in x with values from matched rows in y and add joins columns from y not present in x. Calls rows_upsert and then full_join.

if design_names is omitted its value is taken from the design_names attribute of the simulation results.

If descriptive_regex is given, columns matching the regular expression in both datasets are compared, a warning is given, if the values of those columns do not match. This is intended to compare descriptive statistics or results of unchanged analysis methods to ensure, that both results stem from an exact replication of the simulation results.

Functions

upsert_merge(): Update or add Rows and Columns

Examples

a <- data.frame(x=5:2, y=5:2, a=5:2)
b <- data.frame(x=1:4, y=1:4+10, b=1:4*10)
upsert_merge(a, b, by="x")
#>   x  y  a  b
#> 1 5  5  5 NA
#> 2 4 14  4 40
#> 3 3 13  3 30
#> 4 2 12  2 20
#> 5 1 11 NA 10
if (FALSE) {
old <- readRDS("old_results.Rds")
new <- readRDS("new_results.Rds")
combined <- merge_additional_results(old, new)
}