Functions for Plotting and Reporting Results

Functions for Plotting and Reporting Results

Usage

results_pivot_longer(data, exclude_from_methods = c("descriptive"))

combined_plot(
  data,
  methods,
  xvars,
  yvar,
  facet_x_vars = c(),
  facet_y_vars = c(),
  split_var = 1,
  heights_plots = c(3, 1),
  scale_stairs = 0.75,
  grid_level = 2,
  scales = "fixed",
  hlines = numeric(0),
  use_colours = NULL,
  use_shapes = NULL
)

Arguments

data

for combined_plto simulation results in long format, as returned by results_pivot_longer.

exclude_from_methods

"methods" that should not be pivoted into long format

methods

methods to include in the plot

xvars

orderd vector of variable names to display on the x axis

yvar

variable name of the variable to be displayed on the y axis (metric)

facet_x_vars

vector of variable names to create columns of facets

facet_y_vars

vector of variable names to create rows of facets

split_var

index of xvars along groups of which the plot should be split

heights_plots

relative heights of the main plot and the stairs on the bottom

scale_stairs

height of the stairs for each variable between 0 and 1

grid_level

depth of loops for which the grid-lines are drawn

scales

passed on to facet_grid

hlines

position of horizontal lines, passed as yintercept to geom_hline

use_colours

optional named vector of colours used in scale_colour_manual

use_shapes

optional named vector of shapes used in scale_shape_manual

Value

dataset in long format with one row per method and scenario and one column per metric

a ggplot/patchwork object conatining the plots

Details

With exclude_from_methods descriptive statistics or results of reference methods can be kept as own columns and used like the columns of the simulation parameters.

use_colours and use_shapes both use the method variable in their respective aesthetics.

Functions

results_pivot_longer(): pivot simulation results into long format

combined_plot(): Nested Loop Plot with optional Facets

Examples

if (FALSE) {
plot_data <- simulation_results |>
  results_pivot_longer()
}
if (FALSE) {
# plot the rejection rate of two methods
combined_plot(
  results_long,
  c("logrank", "max_combo"),
  c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"),
  "rejection_0.025",
  grid_level=2
)

# use custom colour and shape scales
# this can be used to group methods by shape or colour
# this is also helpful if methods should have the same aesthetics across plots
my_colours <- c(
  logrank="black",
  max_combo="blue",
  fh_1_0="green",
  fh_0_1="red"
)

my_shapes <- c(
  logrank=1,
  max_combo=2,
  fh_1_0=3,
  fh_0_1=3
)

combined_plot(
  results_long,
  c("logrank", "max_combo", "fh_1_0", "fh_0_1"),
  c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"),
  "rejection_0.025",
  grid_level=2,
  use_colours = my_colours,
  use_shapes = my_shapes
)

combined_plot(
  results_long,
  c("logrank", "max_combo"),
  c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"),
  "rejection_0.025",
  grid_level=2,
  use_colours = my_colours,
  use_shapes = my_shapes
)

# if one has a dataset of metadata with categories of methods
# one could uses those two definitions
# colours for methods, same shapes for methods of same category
my_colours <- ggplot2::scale_colour_discrete()$palette(n=nrow(metadata)) |>
  sample() |>
  setNames(metadata$method)

my_shapes <- metadata$category |>
  as.factor() |>
  as.integer() |>
  setNames(metadata$method)

  combined_plot(
results_long,
c("logrank", "max_combo", "fh_1_0", "fh_0_1"),
c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"),
"rejection_0.025",
grid_level=2,
use_colours = my_colours,
use_shapes = my_shapes
)

combined_plot(
  results_long,
  c("logrank", "max_combo"),
  c("effect_size_ph", "delay", "hazard_ctrl", "n_pat_design", "recruitment", "censoring_prop"),
  "rejection_0.025",
  grid_level=2,
  use_colours = my_colours,
  use_shapes = my_shapes
)
}