Create a Function for Descriptive Statistics of a Dataset

Create a Function for Descriptive Statistics of a Dataset

Usage

analyse_describe()

summarise_describe(name = NULL)

Arguments

name

name for the summarise function, appended to the name of the analysis method in the final results

Value

an analyse function that returns a list with the elements

followup follow up time

events table of events vs. treatment

ice if column ice is present, table of intercurrent events, events, treatment

subgroup if column subgroup is present, table of subgroup, events, treatment

A function that can be used in Summarise that returns a data frame with columns with means and standard deviations for every variable in the description.

Functions

summarise_describe(): Summarise Descriptive Statistics

Examples

condition <- merge(
    assumptions_delayed_effect(),
    design_fixed_followup(),
    by=NULL
  ) |>
  head(1)
#> expand.grid(
#>   delay=seq(0, 10, by=2), # delay of 0, 1, ..., 10 days
#>   hazard_ctrl=0.2,        # hazard under control and before treatment effect
#>   hazard_trt=0.02,        # hazard after onset of treatment effect
#>   random_withdrawal=0.01  # rate of random withdrawal
#> )
#> expand.grid(
#>   n_trt=50,      # 100 patients in the treatment arm
#>   n_ctrl=50,     # 100 patients in the control arm
#>   followup=200,  # fixed followup time
#>   recruitment=50 # recruitment time
#> 
#> )
dat <- generate_delayed_effect(condition)
analyse_describe()(condition, dat)
#> $n_pat
#> [1] 100
#> 
#> $max_followup
#> [1] 135
#> 
#> $evt
#> [1] 100
#> 
#> $evt_ctrl
#> [1] 50
#> 
#> $evt_trt
#> [1] 50
#> 
#> $study_time
#> [1] NA
#> 
condition <- merge(
  assumptions_delayed_effect(),
  design_fixed_followup(),
  by=NULL
) |>
  tail(4) |>
  head(1)
#> expand.grid(
#>   delay=seq(0, 10, by=2), # delay of 0, 1, ..., 10 days
#>   hazard_ctrl=0.2,        # hazard under control and before treatment effect
#>   hazard_trt=0.02,        # hazard after onset of treatment effect
#>   random_withdrawal=0.01  # rate of random withdrawal
#> )
#> expand.grid(
#>   n_trt=50,      # 100 patients in the treatment arm
#>   n_ctrl=50,     # 100 patients in the control arm
#>   followup=200,  # fixed followup time
#>   recruitment=50 # recruitment time
#> 
#> )

summarise_all <- create_summarise_function(
  describe=summarise_describe()
)

# runs simulations
sim_results <- runSimulation(
  design=condition,
  replications=100,
  generate=generate_delayed_effect,
  analyse=list(
    describe=analyse_describe()
  ),
  summarise = summarise_all
)
#> 
#> Design row: 1/1;   RAM used: 159.7 Mb;   Total elapsed time: 0.00s 
#>  Conditions: delay=4, hzrd_c=0.2, hzrd_t=0.02, rndm_w=0.01, n_trt=50, n_ctrl=50, follwp=200, rcrtmn=50
#> 
#> 
#> Simulation complete. Total execution time: 0.12s

# study time is missing, since there was no admin. censoring
sim_results[, 9:16]
#> # A tibble: 1 × 8
#>   describe.n_pat describe.max_followup describe.evt describe.evt_ctrl
#>            <dbl>                 <dbl>        <dbl>             <dbl>
#> 1            100                 190.1          100                50
#> # ℹ 4 more variables: describe.evt_trt <dbl>, describe.study_time <dbl>,
#> #   describe.sd_n_pat <dbl>, describe.sd_max_followup <dbl>