Create Analyse function for piecewise exponential model

Create Analyse function for piecewise exponential model

Usage

analyse_piecewise_exponential(cuts, testing_only = FALSE)

Arguments

cuts

interval boundaries for the piecewise exponential model

testing_only

if set to TRUE omits all statistics in the intervals and just returns the p value of the global test.

Value

an analyse function that can be used in runSimulation

Details

If there's any time interval no patients ever enter, NA is returned for all time intervals. This behavior will likely change in future package versions.

Examples

condition <- merge(
   assumptions_delayed_effect(),
   design_fixed_followup(),
   by=NULL
 ) |>
 head(1)
#> expand.grid(
#>   delay=m2d(seq(0, 10, by=2)), # delay of 0, 1, ..., 10 months
#>   hazard_ctrl=m2r(24),         # median survival control of 24 months
#>   hazard_trt=m2r(36),          # median survival treatment of 36 months
#>   random_withdrawal=m2r(120)   # median time to random withdrawal 10 years
#> )
#> expand.grid(
#>   n_trt=150,         # 150 patients in the treatment arm
#>   n_ctrl=150,        # 150 patients in the control arm
#>   followup=m2d(24),  # followup 2 years
#>   recruitment=m2d(6) # recruitment time 6 months
#> 
#> )
dat <- generate_delayed_effect(condition)
analyse_piecewise_exponential(cuts=c(90, 360))(condition, dat)
#> $p
#> [1] 0.001247528
#> 
#> $hr_s
#> # A tibble: 1 × 3
#>   `trt:intervalI1` `trt:intervalI2` `trt:intervalI3`
#>              <dbl>            <dbl>            <dbl>
#> 1            0.723            0.662            0.617
#> 
#> $p_s
#> # A tibble: 1 × 3
#>   `trt:intervalI1` `trt:intervalI2` `trt:intervalI3`
#>              <dbl>            <dbl>            <dbl>
#> 1            0.486            0.148         0.000250
#> 
#> $hr_s_lower
#> # A tibble: 1 × 3
#>   `trt:intervalI1` `trt:intervalI2` `trt:intervalI3`
#>              <dbl>            <dbl>            <dbl>
#> 1            -1.27           -0.983           -0.742
#> 
#> $hr_s_upper
#> # A tibble: 1 × 3
#>   `trt:intervalI1` `trt:intervalI2` `trt:intervalI3`
#>              <dbl>            <dbl>            <dbl>
#> 1            0.581            0.140           -0.224
#> 
#> $N_pat
#> [1] 300
#> 
#> $N_evt
#> [1] 300
#> 
#> $interval_table
#>     
#>        0   1
#>   I1 281  19
#>   I2 230  51
#>   I3   0 230
#>